requirements.txt
google-auth==2.39.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.167.0
flask==3.1.0
flask-wtf==1.2.2
flask-bootstrap==3.3.7.1
pandas==2.2.3
matplotlib==3.10.1
seaborn==0.13.2
reportlab==4.4.0
googletrans==4.0.0-rc1
pdfkit==1.0.0
python-dotenv==1.0.0
